Add cured hide patches, leather tarp raincatch
* Added a fur/leather patch sized version of cured hides and pelts,
accordingly using a fifth of the material. Also includes uncrafts and
related recipes.
* Added recipe overrides to tanning leather and fur, to allow using an
appropriate amount of the smaller hide. I'd add a version that tans
directly into leather or fur patches, but wasn't sure how to split up
the materials for the recipe. It's not as easily split as salt-curing.
* Changed the fur/leather wrap overrides to use patches of cured
hide/pelt in place of whole chunks of full-size cured hides/pelts.
Likewise applied the change to the additional wrap items the mod adds.
* Added a leather tarp version of the tarp raincatcher. Uses cordage
instead of ONLY strings given it's likely to be made in wilderness
scenarios.