Various changes, couple additions
* Minor item idea: makeshift gas hoods as a step in between the filter
mask and gas mask, based on the concept of the hypo helmet and similar
WWI-era masks. Easier to craft than a survivor mask, but a legit gas
mask might be worthwhile if you have the recipe and resources. Technical
issues mean the item still has to be activated first, but once activated
it can't be turned off (as expected, the whole thing is basically a
wearable filter), and it's effectively disposable.
* Tweaked volume capacity of placed hobo stove so you have to use fuel
smaller than sticks or planks.
* Made it so that knocking over a hobo stove just returns the item
rather than destroying it, but in exchange lowered bash threshold.
* Dummied out small cured pelts/hides, as now vanilla cured pelts and
hides are roughly patch-sized. Replaced their use in recipes with these
new patch-sized pelts and hides were needed. Scraps of tanning
fur/leather have been retained however, since the normal tanning recipes
take 6 cured pelts/hides at a time.
* Fixed stone spears not being autolearned at a reasonable level.
* Lowered skill demand for making tar oil a bit.
* Pushed the spawns of feral livestock back heavily, more in line with
the fact that 91-day-seasons are default.
* Couple typo fixes.