QoL ideas
* Set it so that rubber slugs and limestone can be used with slings. The
former has a slight chance of survving impact and will stun targets,
while the latter can blind your target.
* Set it so that petrified eyes, spiral stones, chunks of sulfur, golf
balls, pool balls, baseballs, and magic 8-balls can be used with staff
slings. Most of these are just for fun as they tend not to hit as hard
as the real deal and are more likely to break, though many of the more
aerodynamic options have marginally greater range, and sulfur chunks
also have the utility of blinding targets.
* Added override so that the "only exists to filter out filth rags"
recipe for makeshift bandages can be made in the darkness.