Consume mod
Incorporated all standard More Survival Tools content into MST Extra,
updated modinfo to relfect that this is now a sequel instead of an addon
to MST. Also a few content fixes and updates here and there.

Will be relevent if they pull the trigger on MST during the final
version of The Unmoddening.

Also gonna need to add a replacement to resinous cord, since they axed
it. Since basic repair kits and duct tape are pretty hard to come by
innawoods, and the niche resinous cord was meant to fill was left empty,
I'm gonna likely have to add makeshift equivalents to both...