Tree bacon
Introduce a method of using pine trees for something that's realistic
AND good for gameplay.

If we ever get JSONized multi-stage harvesting, I plan to rework it so
that you have to opt to harvest a harvested pine tree, which will kill
the tree and make inner bark less renewable. But since
double-harvestable terrain isn't a thing yet, enjoy a windfall until
then.