Fix quick butchery/dissection, missing tool quality
* Added overrides to the qualities introduced by
https://github.com/CleverRaven/Cataclysm-DDA/pull/42372 to fix quick
butchery needing a surface, because that is an astoundingly tedious
idea.
* Also reduced surface demands for dissection so that low-level surfaces
work for large creatures, and small creatures don't need any surface.
* This also fixes dismembering them, I'm not even sure what absurdity in
the code is tying dismemberment requirement to one of the other butchery
requirements, but needing a clean fancy tarp to messily hack apart a
body is a terrible idea.
* Added an override to tarps, adding surface qaulity because
https://github.com/CleverRaven/Cataclysm-DDA/pull/42553 forgot them.
* Cleaned out an obsolete item that hasn't been used since back when I
was still a contributor. It was pinging the zero-weight and there was
really no need to update it after literally 3-4 years of it being
dummied out.
* Zero-weight fixes for psuedo-items.