Fix stone chopper recipe, misc JSON updates
* Added an override for the stone chopper, lowering fabrication to one
and making it slower withouit knapping proficiency. I'll retain this in
case https://github.com/CleverRaven/Cataclysm-DDA/pull/42740 gets killed
for being realisitic in a way that benefits the player. If they force
knapping to be required for what's basically a pebble tool that even I
could whip up in about a minute IRL, I'll definitely keep the override
in.
* Removed the override for makeshift bandages since they eventually made
it BLIND_EASY.
* Added brains to the options for tanning small patches of hide/pelt,
since the small-scale recipe was rather short on good options that scale
down enough to use.
* Obsoleted the now-redundant recipe for stone spears since the base
recipe got un-obsoleted in vanilla. It may still end up needing an
override instead if the recipe stays prohibtive, but making it require
will likely be acceptable now that you can actually learn proficiencies.
* Updated tool, material, and proficency needs for stone javelins, flint
and steel, sharp rock, and makeshift hand drill with the knapping update
in mind.