Updates to makeshift gas hood
* Removed the flags that were previously needed for making the mask
dispoable, now that you have to explicitly load it with filter charges
it should be loadable. Thanks to Tiny-wait on the reddit for spotting
that.
* Also bumped up the EP value because unhardcoded toxic gas no longer
seems to work very well the way it used to. Now it seems that resistance
to toxic gas is binary, either you have 15 EP, or you don't.