Potted meat, minor DDA fixes
Added potted meat, also known as confit. Rather than corned beef or whatever one might think of, this refers to a historical style of food preservation where butter (the norm with potted meat) or fat (common with confit) is used to immerse and seal meat, fish, and the like.

Item's set to last comparably long to pemmican, and copies from it in lieu of any other good options since that's needed for compatibility with Nonperishable Overhaul. Can be extremely calorie-dense if you go sealing every chunk of meat you get in a glob of lard.

Recipe has difficulty on par with making jerky, but autolearns a level higher since it is more involved. Family cookbooks and the Scots cookbook can teach it sooner.