Add license
This mod is a derivative mod work of a mod that lived in the DDA repo,
and so is licensed under the CC-BY-SA 3.0 license, as DDA is.
Introduced to the DDA repo in
https://github.com/CleverRaven/Cataclysm-DDA/pull/11707

Copy the license from DDA, and make a few changes to make it relevant.

Bright Nights is also licensed under CC-BY-SA 3.0, so this will work for
both versions.