Remove clay lamp oil from DDA version
I didn't notice that https://github.com/CleverRaven/Cataclysm-DDA/pull/55596 ported this over to mainline DDA, since the ID matches I can afford to remove this from the DDA version instead of just obsoleting it.

Closes https://github.com/chaosvolt/MST_Extra_Mod/issues/35

Note to self: I need to also update the DDA version of Arcana, as I'd forgot that itemgroup reading_lights is actually a group in the mod, not a vanilla itemgroup.